fix(update): init app before loading other modules
